LayoutTests/imported/w3c/web-platform-tests/webcodecs/videoDecoder-codec-specific.https.any_h264_avc-expected.txt


PASS Test isConfigSupported()
PASS Test isConfigSupported() with 1080p crop
PASS Test that isConfigSupported() returns a parsed configuration
PASS Test invalid configs
FAIL Test configure() assert_equals: state expected "configured" but got "closed"
FAIL Decode a key frame promise_test: Unhandled rejection with value: object "InvalidStateError: VideoDecoder is not configured"
FAIL Decode a non key frame first fails assert_throws_dom: function "() => decoder.decode(CHUNKS[1], 'decode')" threw object "InvalidStateError: VideoDecoder is not configured" that is not a DOMException DataError: property "code" is equal to 11, expected 0
FAIL Verify reset() suppresses outputs promise_test: Unhandled rejection with value: object "InvalidStateError: VideoDecoder is not configured"
PASS Test unconfigured VideoDecoder operations
PASS Test closed VideoDecoder operations
FAIL Decode empty frame promise_test: Unhandled rejection with value: object "InvalidStateError: VideoDecoder is not configured"
FAIL Decode corrupt frame promise_test: Unhandled rejection with value: object "InvalidStateError: VideoDecoder is not configured"
FAIL Close while decoding corrupt frame promise_test: Unhandled rejection with value: object "InvalidStateError: VideoDecoder is not configured"
FAIL Test decoding after flush promise_test: Unhandled rejection with value: object "InvalidStateError: VideoDecoder is not configured"
FAIL Test decoding a with negative timestamp promise_test: Unhandled rejection with value: object "InvalidStateError: VideoDecoder is not configured"
FAIL Test reset during flush promise_test: Unhandled rejection with value: object "InvalidStateError: VideoDecoder is not configured"
FAIL Test low-latency decoding promise_test: Unhandled rejection with value: object "InvalidStateError: VideoDecoder is not configured"
FAIL VideoDecoder decodeQueueSize test promise_test: Unhandled rejection with value: object "InvalidStateError: VideoDecoder is not configured"
